+++ packages/cryptography/test/unit/encoding/der.js
@@ -68,8 +68,7 @@ describe("encoding/der", function () {
                                                     seq: [
                                                         // HMAC-SHA-256
                                                         {
-                                                            ident:
-                                                                "1.2.840.113549.2.9",
+                                                            ident: "1.2.840.113549.2.9",
                                                         },
                                                         // no parameters
                                                         {},